helios: Fix build under x86_64 Haiku
* Drop BExperimental color_tools for stable Haiku API
* Fix 64-bit type build errors

+++ includes/BubbleHelper.h
@@ -51,7 +51,7 @@ class BubbleHelper
 		void HideBubble();
 		void ShowBubble(BPoint dest);
 		
-		static long runcount;
+		static int32 runcount;
 };
 
 #endif
